Clamp Light code completed, needs to be tested.
Created a method that sets a flag after testing if the current max is
reached by the clamp motor. Also created a method that tests the flag
and turns the light on if it is true, it is ran at the end of execute.
Also removed the starting calibration work to the elevator and will
create a new branch for that off of tuning.